requirements/base.txt
django==1.8.1
django-environ==0.3.0
pyswisseph==2.00.00-2
geocoder==1.3.0
pytz==2015.2
djangorestframework==3.1.2
python-dateutil==2.4.2
pyOpenSSL==0.15.1
ndg-httpsclient==0.4.0
pyasn1==0.1.7
